requirements.txt
git+git://github.com/lekhakpadmanabh/Summarizer.git
goose-extractor==1.0.22
praw==3.0.0
git+git://github.com/networkx/networkx.git
numpy==1.8.1
scipy==0.14.0
scikit-learn==0.15.2
sumy==0.3.0